// src/enums/enum.ts
export enum ReportCategory {
  HATE = "hate",
  ABUSE_HARASSMENT = "abuse_harassment",
  VIOLENT_SPEECH = "violent_speech",
  SPAM = "spam",
  PRIVACY = "privacy",
  OTHERS = "others",
}

export enum PoliticalParty {
  PKB = "Partai Kebangkitan Bangsa",
  GERINDRA = "Partai Gerindra",
  PDIP = "Partai Demokrasi Indonesia Perjuangan",
  GOLKAR = "Partai Golkar",
  NASDEM = "Partai Nasional Demokrat",
  BURUH = "Partai Buruh",
  GELORA = "Partai Gelora",
  PKS = "Partai Keadilan Sejahtera",
  PKN = "Partai Kebangkitan Nusantara",
  HANURA = "Partai Hanura",
  GARUDA = "Partai Garuda",
  PAN = "Partai Amanat Nasional",
  PBB = "Partai Bulan Bintang",
  DEMOKRAT = "Partai Demokrat",
  PSI = "Partai Solidaritas Indonesia",
  PERINDO = "Partai Perindo",
  PPP = "Partai Persatuan Pembangunan",
  PU = "Partai Ummat",
}

export enum NIKVerificationStatus {
  DRAFT = "draft",
  PENDING = "pending",
  VERIFIED = "verified",
  REJECTED = "rejected",
}